Increase unreachable bridge state timeout to 30s
In the situation where network is having issues and a bridge is rapidly
reconnecting to asmux over and over, it's possible that things get into
a bad state where the unreachable state is sent after a connected state.

This doesn't fix that issue but it should reduce the chances of it
occuring significantly. A full fix may require locking on the per bridge
or more complex state management for asmux.